+++ R/weight_draws.R
@@ -14,6 +14,8 @@
 #' @param log (logical) Are the weights passed already on the log scale? The
 #'   default is `FALSE`, that is, expecting `weights` to be on the standard
 #'   (non-log) scale.
+#' @param pareto_smooth (logical) Should the weights be Pareto-smoothed?
+#' The default is `FALSE`.
 #' @template args-methods-dots
 #' @template return-draws
 #'
@@ -43,6 +45,9 @@
 #' head(weights(x))
 #' head(weights(x, log=TRUE, normalize = FALSE)) # recover original log_wts
 #'
+#' # add weights on log scale and Pareto smooth them
+#' x <- weight_draws(x, weights = log_wts, log = TRUE, pareto_smooth = TRUE)
+#'
 #' @export
 weight_draws <- function(x, weights, ...) {
   UseMethod("weight_draws")
@@ -50,9 +55,15 @@ weight_draws <- function(x, weights, ...) {
 
 #' @rdname weight_draws
 #' @export
-weight_draws.draws_matrix <- function(x, weights, log = FALSE, ...) {
+weight_draws.draws_matrix <- function(x, weights, log = FALSE, pareto_smooth = FALSE, ...) {
+
+
+  pareto_smooth <- as_one_logical(pareto_smooth)
   log <- as_one_logical(log)
   log_weights <- validate_weights(weights, x, log = log)
+  if (pareto_smooth) {
+    log_weights <- pareto_smooth_log_weights(log_weights)
+  }
   if (".log_weight" %in% variables(x, reserved = TRUE)) {
     # overwrite existing weights
     x[, ".log_weight"] <- log_weights
@@ -66,9 +77,14 @@ weight_draws.draws_matrix <- function(x, weights, log = FALSE, ...) {
 
 #' @rdname weight_draws
 #' @export
-weight_draws.draws_array <- function(x, weights, log = FALSE, ...) {
+weight_draws.draws_array <- function(x, weights, log = FALSE, pareto_smooth = FALSE, ...) {
+
+  pareto_smooth <- as_one_logical(pareto_smooth)
   log <- as_one_logical(log)
   log_weights <- validate_weights(weights, x, log = log)
+  if (pareto_smooth) {
+    log_weights <- pareto_smooth_log_weights(log_weights)
+  }
   if (".log_weight" %in% variables(x, reserved = TRUE)) {
     # overwrite existing weights
     x[, , ".log_weight"] <- log_weights
@@ -82,18 +98,28 @@ weight_draws.draws_array <- function(x, weights, log = FALSE, ...) {
 
 #' @rdname weight_draws
 #' @export
-weight_draws.draws_df <- function(x, weights, log = FALSE, ...) {
+weight_draws.draws_df <- function(x, weights, log = FALSE, pareto_smooth = FALSE, ...) {
+
+  pareto_smooth <- as_one_logical(pareto_smooth)
   log <- as_one_logical(log)
   log_weights <- validate_weights(weights, x, log = log)
+  if (pareto_smooth) {
+    log_weights <- pareto_smooth_log_weights(log_weights)
+  }
   x$.log_weight <- log_weights
   x
 }
 
 #' @rdname weight_draws
 #' @export
-weight_draws.draws_list <- function(x, weights, log = FALSE, ...) {
+weight_draws.draws_list <- function(x, weights, log = FALSE, pareto_smooth = FALSE, ...) {
+
+  pareto_smooth <- as_one_logical(pareto_smooth)
   log <- as_one_logical(log)
   log_weights <- validate_weights(weights, x, log = log)
+  if (pareto_smooth) {
+    log_weights <- pareto_smooth_log_weights(log_weights)
+  }
   niterations <- niterations(x)
   for (i in seq_len(nchains(x))) {
     sel <- (1 + (i - 1) * niterations):(i * niterations)
@@ -104,9 +130,14 @@ weight_draws.draws_list <- function(x, weights, log = FALSE, ...) {
 
 #' @rdname weight_draws
 #' @export
-weight_draws.draws_rvars <- function(x, weights, log = FALSE, ...) {
+weight_draws.draws_rvars <- function(x, weights, log = FALSE, pareto_smooth = FALSE, ...) {
+
+  pareto_smooth <- as_one_logical(pareto_smooth)
   log <- as_one_logical(log)
   log_weights <- validate_weights(weights, x, log = log)
+  if (pareto_smooth) {
+    log_weights <- pareto_smooth_log_weights(log_weights)
+  }
   x$.log_weight <- rvar(log_weights)
   x
 }
@@ -160,4 +191,15 @@ validate_weights <- function(weights, draws, log = FALSE) {
     weights <- log(weights)
   }
   weights
+}
+
+
+pareto_smooth_log_weights <- function(log_weights) {
+  pareto_smooth(
+    log_weights,
+    tail = "right",
+    return_k = TRUE,
+    are_log_weights = TRUE,
+    extra_diags = TRUE
+  )$x
 }