+++ man-roxygen/ref-vehtari-paretosmooth-2022.R
@@ -1,4 +1,4 @@
 #' @references
 #' Aki Vehtari, Daniel Simpson, Andrew Gelman, Yuling Yao and
 #' Jonah Gabry (2022). Pareto Smoothed Importance Sampling.
-#' arxiv:arXiv:1507.02646
+#' arxiv:arXiv:1507.02646 (version 8)